% Auto-generated: do not edit by hand
\name{autoLayout}

\alias{autoLayout}

\title{AutoLayout component}

\description{

}

\usage{
autoLayout(children=NULL, id=NULL, direction=NULL, spacing=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The content of the layout.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{direction}{A value equal to: 'row', 'row-reverse', 'column', 'column-reverse'. Defines the `flex-direction` style property.}

\item{spacing}{Numeric. Defines the space between the type `item` components.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

