% Auto-generated: do not edit by hand
\name{calculator}

\alias{calculator}

\title{Calculator component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
calculator(id=NULL, result=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{result}{Character. The result of the calculation.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

