% Auto-generated: do not edit by hand
\name{card}

\alias{card}

\title{Card component}

\description{

}

\usage{
card(id=NULL, buttonText=NULL, mainContent=NULL, n_clicks=NULL,
secondaryContent=NULL, subtitle=NULL, title=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{buttonText}{Character. The text for the button.}

\item{mainContent}{Character. The main content of the card.}

\item{n_clicks}{Numeric. Number of times the button has been clicked.}

\item{secondaryContent}{Character. The secondary content of the card.}

\item{subtitle}{Character. The subtitle of the card.}

\item{title}{Character. The title of the card.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

